GATES Estimation

Estimates the group average treatment effects, with the groups formed by cutting the distribution of the estimated CATEs into K quantiles.

blp_regression(y, D, cates, pscore, mu, mu0, mu1, scores)

Arguments

y

Observed outcomes.

D

Treatment indicator.

cates

Estimated CATEs. CATEs must be estimated with different observations than those in y and D.

pscore

Propensity scores. If unknown, they must be estimated using different observations than those in y and D.

mu

Estimated regression function. It must be estimated with different observations than those in y and D.

mu0

Estimated regression function for control units. It must be estimated with different observations than those in y and D.

mu1

Estimated regression function for treated units. It must be estimated with different observations than those in y and D.

scores

Estimated doubly-robust scores. They must be estimated via K-fold cross-fitting.

Value

Return.

Details

Currently, the routine supports only single-splits.
We rearrange the estimate GATES to obey the monotonicity property.

See also

Author

Riccardo Di Francesco

Examples

## Generate data.
set.seed(1986)

n <- 3000
k <- 3