CATEs Evaluation

Evaluates the quality of CATEs estimates by implementing several Generic Machine Learning methodologies.

evalue_cates(y, D, X, cates, is_train, pscore = NULL, verbose = TRUE)

Arguments

y

Observed outcomes.

D

Treatment indicator.

X

Covariate matrix (no intercept).

cates

Estimated CATEs. CATEs must be estimated using only the training sample.

is_train

Logical vector denoting which observations belong to the training sample.

pscore

Propensity scores. If unknown, they must be estimated using only the training sample. If not provided by the user, they are estimated internally via an honest regression_forest.

verbose

Logical, set to FALSE to prevent the function from printing the progresses.

Value

Return.

Details

Currently, the routine supports only single-splits.
Regression functions are estimated via honest regression_forests.
If strategy is set to "aipw", doubly-robust scores are estimated internally using the validation sample via 5-fold cross fitting and using honest regression forests (see the dr_scores function).
If not provided, cates are estimated internally using a causal forest estimator and the training sample.
If not provided, pscore are estimated internally using a regression forest.

See also

Author

Riccardo Di Francesco

Examples

## Generate data.
set.seed(1986)

n <- 3000
k <- 3